working split of remote and docker
