+++ src/rust/engine/process_execution/remote/Cargo.toml
@@ -1,0 +1,52 @@
+[package]
+version = "0.0.1"
+edition = "2021"
+name = "remote"
+authors = [ "Pants Build <pantsbuild@gmail.com>" ]
+publish = false
+
+[dependencies]
+async-stream = "0.3"
+async-trait = "0.1"
+async-lock = "2.5"
+protos = { path = "../../protos" }
+bytes = "1.2"
+derivative = "2.2"
+# TODO: Waiting on https://github.com/Aeledfyr/deepsize/pull/{30,31,32}.
+deepsize = { git = "https://github.com/stuhood/deepsize.git", rev = "5c8bee5443fcafe4aaa9274490d354412d0955c1", features=["log"] }
+grpc_util = { path = "../../grpc_util" }
+fs = { path = "../../fs" }
+futures = "0.3"
+hashing = { path = "../../hashing" }
+libc = "0.2.137"
+log = "0.4"
+store = { path = "../../fs/store" }
+task_executor = { path = "../../task_executor" }
+concrete_time = { path = "../../concrete_time" }
+tokio = { version = "1.21", features = ["net", "process", "rt-multi-thread", "sync", "time"] }
+tokio-rustls = "0.23"
+tokio-util = { version = "0.7", features = ["codec"] }
+uuid = { version = "1.1", features = ["v4"] }
+workunit_store = { path = "../../workunit_store" }
+async-oncecell = "0.2"
+once_cell = "1.15"
+prost = "0.9"
+prost-types = "0.9"
+rand = "0.8"
+tonic = { version = "0.6", features = ["transport", "codegen", "tls", "tls-roots", "prost"] }
+tryfuture = { path = "../../tryfuture" }
+process_execution = { path = ".." }
+strum = "0.24"
+strum_macros = "0.24"
+parking_lot = "0.12"
+
+[dev-dependencies]
+env_logger = "0.9.0"
+maplit = "1.0.1"
+mock = { path = "../../testutil/mock" }
+parking_lot = "0.12"
+sharded_lmdb = { path = "../../sharded_lmdb" }
+spectral = "0.6.0"
+tempfile = "3"
+testutil = { path = "../../testutil" }
+tokio = { version = "1.21", features = ["macros"] }
