+++ src/rust/engine/process_executor/src/main.rs
@@ -297,7 +297,7 @@ async fn main() {
         );
       }
 
-      let remote_runner = process_execution::remote::CommandRunner::new(
+      let remote_runner = remote::remote::CommandRunner::new(
         &address,
         process_metadata.instance_name.clone(),
         process_metadata.cache_key_gen_version.clone(),
@@ -315,7 +315,7 @@ async fn main() {
 
       let command_runner_box: Box<dyn process_execution::CommandRunner> = {
         Box::new(
-          process_execution::remote_cache::CommandRunner::new(
+          remote::remote_cache::CommandRunner::new(
             Arc::new(remote_runner),
             process_metadata.instance_name.clone(),
             process_metadata.cache_key_gen_version.clone(),
@@ -326,7 +326,7 @@ async fn main() {
             headers,
             true,
             true,
-            process_execution::remote_cache::RemoteCacheWarningsBehavior::Backoff,
+            remote::remote_cache::RemoteCacheWarningsBehavior::Backoff,
             CacheContentBehavior::Defer,
             args.cache_rpc_concurrency,
             Duration::from_secs(2),
@@ -563,7 +563,7 @@ async fn extract_request_from_action_digest(
       .filter(|env| {
         // Filter out environment variables which will be (re-)set by ExecutionRequest
         // construction.
-        env.name != process_execution::remote::CACHE_KEY_TARGET_PLATFORM_ENV_VAR_NAME
+        env.name != process_execution::CACHE_KEY_TARGET_PLATFORM_ENV_VAR_NAME
       })
       .map(|env| (env.name.clone(), env.value.clone()))
       .collect(),
